company_research_agent_project/requirements.txt
streamlit
langchain
langchain-community
langchain-together
beautifulsoup4
tavily-python
python-dotenv  # For loading .env file